requirements.txt
# MCDA Core - Python Dependencies
# 多准则决策分析核心框架依赖包

# 核心计算
numpy>=2.4.0
scipy>=1.17.0

# 数据处理
PyYAML>=6.0
tabulate>=0.9.0

# 测试框架
pytest>=9.0.0
pytest-cov>=5.0.0

# 代码质量（可选）
# black>=24.0.0
# flake8>=7.0.0
# mypy>=1.8.0
openpyxl==3.1.5
pandas==3.0.0
python-dateutil==2.9.0.post0
six==1.17.0
